test(daemon): strengthen Phase 4 TCB hardening baseline verification + cross-platform support
- Split Linux-only hardening logic (dropCapabilities, applySeccompFilter,
  setResourceLimits, bounding set) into daemon_hardening_linux.go +
  daemon_hardening_other.go (no-op stubs). This follows the existing
  secrets_io_* pattern and makes `go test ./...` and `go build` work on
  non-Linux platforms (Docker sbx path per host-daemon spec).

- Add real verification tests in the new daemon_hardening_linux_test.go:
  - TestHardening_ResourceLimitsActuallyApplied (uses Getrlimit)
  - TestHardening_CapabilitiesBoundingOrEffective (uses capget)
  - TestHardening_SeccompInstallHook (documents the intentional limitation
    that STRICT=1 cannot be exercised inside a normal Go test process)

- Minor updates to the smoke tests in daemon_tcb_test.go and a clarifying
  comment in start.go around the UnixPeerAllow override vs Default policy.

- Update traceability matrix (03-daemon-minimal-tcb-refactor.md) and
  the Phase 4 baseline doc to reflect the improved test coverage.

- Fix pre-existing compile errors in internal/api that completely blocked
  `go test` on this branch:
  - server.go: bad JSON struct tag + unused okUID variable
  - server_unix_policy_linux_test.go: missing parameter name in handler func

- Add entry to CHANGELOG.md.

These changes make the Phase 4 "hardening baseline" claims actually
testable and provide meaningful regression protection for the Host
Daemon TCB (caps, seccomp hook, resource limits, socket creation).
