+++ docs/implementation-plan/03-daemon-minimal-tcb-refactor.md
@@ -46,7 +46,7 @@ The Host Daemon intentionally retains only:
 
 | Requirement | Test intent | Type | Location | CI tier | Status |
 |-------------|-------------|------|----------|---------|--------|
-| Minimal privilege | Capability drop and bounding set invoked without panicking; stricter proofs where CAPs available | Unit / smoke | `cmd/aegisclaw/daemon_tcb_test.go` (`TestHardening_CapabilitiesDropCalled`, `TestHardening_CapabilityBoundingSetApplied`) | PR | **Partial** — environment-dependent; skips/errors logged, not full cap inventory assertion |
+| Minimal privilege | Capability drop and bounding set invoked without panicking; stricter proofs where CAPs available | Unit / smoke | `cmd/aegisclaw/daemon_tcb_test.go` (`TestHardening_CapabilitiesDropCalled`, `TestHardening_CapabilityBoundingSetApplied`) | PR | **Implemented** (with env notes) — smoke in `daemon_tcb_test.go`; real post-drop capget + bounding verification + rlimit Getrlimit asserts in `daemon_hardening_linux_test.go`; non-privileged runs log/skip cleanly |
 | No secret handling | Daemon does not initialize vault/court/build paths as part of minimal runtime init documented for TCB | Unit / structural | `cmd/aegisclaw/daemon_tcb_test.go` (`TestDaemonDoesNotInitializeForbiddenComponents`, `TestNoNonTCBInitializations`) | PR | **Partial** — several tests rely on compile-time removal + weak runtime assertions; strengthen with explicit field/build tags where possible |
 | Keypair isolation | Private keys for Merkle and VM contracts do not leak across packages / wrong verify paths | Unit | `internal/audit/merkle_test.go` (`TestMerkleLog_WrongKeyDetection`, tamper tests); **`internal/kernel/kernel_test.go` (`TestKernelInit_LogsDoNotContainRawPrivateKeyMaterial` — DB-04 slice)** | PR | **Partial** — proves Merkle crypto + kernel init logs; full “never leaves VM” distribution still backlog |
 | Lifecycle containment | On daemon exit, monitors stop cleanly; future: no orphaned VMs | Unit / INT | `cmd/aegisclaw/lifecycle_integration_test.go`; `cmd/aegisclaw/daemon_test.go` (monitor threshold); `cmd/aegisclaw/microvm_shutdown_test.go`; `cmd/aegisclaw/shutdown_all_sandboxes_test.go`; **`daemon_shutdown_sigterm_e2e_test.go` (DB-01 subprocess slice)** | INT / PR | **Partial** — unit + subprocess stub path; full process kill + real Firecracker still [06](06-sandbox-lifecycle-containment.md) |
@@ -61,8 +61,8 @@ The Host Daemon intentionally retains only:
 | Hardening area | Test intent | Type | Location | CI tier | Status |
 |----------------|-------------|------|----------|---------|--------|
 | Lifecycle containment (signals + cleanup) | Hub/Store IDs get Stop+Delete; `List`→Stop/Delete for all sandboxes | Unit | `start.go` `terminateManagedHubAndStoreVMs`; `daemon_handlers_extended.go` `shutdownAllSandboxes`; `microvm_shutdown_test.go`, `shutdown_all_sandboxes_test.go`, **`daemon_shutdown_sigterm_e2e_test.go`** | PR | **Partial** — logic + subprocess SIGTERM slice; full process kill + real VMs still **gap** |
-| Capability dropping | `dropCapabilities` / bounding set exercised | Unit | `daemon_tcb_test.go` | PR | **Partial** |
-| seccomp-bpf | Filter install hook runs | Unit | `TestHardening_SeccompFilterHook` | PR | **Partial** — non-fatal in many envs |
+| Capability dropping | `dropCapabilities` / bounding set exercised | Unit | `daemon_tcb_test.go` | PR | **Implemented** (with env notes) — exercised + verified via new Linux tests (see above rows) |
+| seccomp-bpf | Filter install hook runs | Unit | `TestHardening_SeccompFilterHook` | PR | **Implemented** (with env notes) — exercise hook in `daemon_tcb_test.go`; strict-mode kill proof via subprocess re-exec (blocked syscall → death) in `daemon_hardening_linux_test.go` |
 | Static binary | Same as host-daemon row | Build | `make build-static` | REL | **Implemented** |
 | Unix socket permissions | `0700` runtime dir pattern, `0600` socket ownership helper | Unit | `internal/paths/paths_test.go` (`TestSetRuntimeSocketOwnerUsesOwnerOnlyMode`, runtime dir tests) | PR | **Implemented** for path helpers; end-to-end bind still see 04 |
